pr review fix

+++ persistence/src/main/java/cz/muni/fi/dao/DriverDaoImpl.java
@@ -68,7 +68,6 @@ public class DriverDaoImpl implements DriverDao{
         if (driver.getId() != null) throw new IllegalArgumentException("add: id must be null");
         if (driver.getName() == null) throw new IllegalArgumentException("add: name is null");
         if (driver.getNationality() == null) throw new IllegalArgumentException("add: nationality is null");
-//        if (driver.getSpecialSkill() == null) throw new IllegalArgumentException("add: skill is null");
         if (driver.getSurname() == null) throw new IllegalArgumentException("add: surname is null");
 
         em.persist(driver);
@@ -79,7 +78,6 @@ public class DriverDaoImpl implements DriverDao{
         if (driver.getId() == null) throw new IllegalArgumentException("update: id is null");
         if (driver.getName() == null) throw new IllegalArgumentException("update: name is null");
         if (driver.getNationality() == null) throw new IllegalArgumentException("update: nationality is null");
-//        if (driver.getSpecialSkill() == null) throw new IllegalArgumentException("update: skill is null");
         if (driver.getSurname() == null) throw new IllegalArgumentException("update: surname is null");
 
         em.merge(driver);
@@ -88,6 +86,9 @@ public class DriverDaoImpl implements DriverDao{
     public void deleteDriver(Driver driver) {
         if (driver == null) throw new IllegalArgumentException("delete: driver is null");
         if (driver.getId() == null) throw new IllegalArgumentException("delete: id is null");
+        if (driver.getName() == null) throw new IllegalArgumentException("delete: name is null");
+        if (driver.getNationality() == null) throw new IllegalArgumentException("delete: nationality is null");
+        if (driver.getSurname() == null) throw new IllegalArgumentException("delete: surname is null");
 
         em.remove(em.merge(driver));
     }
